rename all dart files to snake_case
This is dart's convention and we should follow it
