+++ lib/domain/entities/base_obj.dart
