+++ lib/infrastructure/datasources/sqlite/sqlite_tables.dart
@@ -1,4 +1,4 @@
-import 'package:cobble/domain/entities/PebbleDevice.dart';
+import 'package:cobble/domain/entities/pebble_device.dart';
 
 final Map<Type, String> typeToTable = {
   PebbleDevice: TableNames.pebbledevice,
