+++ lib/main.dart
@@ -1,14 +1,14 @@
 import 'dart:ui';
 
-import 'package:cobble/ui/Localize.dart';
-import 'package:cobble/ui/Router.dart' as router;
-import 'package:cobble/ui/Theme.dart';
+import 'package:cobble/ui/localize.dart';
+import 'package:cobble/ui/router.dart' as router;
+import 'package:cobble/ui/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:hooks_riverpod/all.dart';
 
-import 'domain/calendar/CalendarPermission.dart';
+import 'domain/calendar/calendar_permission.dart';
 
 String getBootUrl = "https://boot.rebble.io/";
