+++ lib/ui/common/icons/comp_icon.dart
