+++ lib/ui/home/tabs/test_tab.dart
@@ -1,7 +1,7 @@
-import 'package:cobble/domain/calendar/CalendarList.dart';
-import 'package:cobble/domain/connection/ConnectionStateProvider.dart';
+import 'package:cobble/domain/calendar/calendar_list.dart';
+import 'package:cobble/domain/connection/connection_state_provider.dart';
 import 'package:cobble/infrastructure/pigeons/pigeons.dart';
-import 'package:cobble/ui/common/icons/fonts/RebbleIconsStroke.dart';
+import 'package:cobble/ui/common/icons/fonts/rebble_icons_stroke.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/all.dart';
