+++ lib/ui/setup/boot/rebble_setup_success.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:cobble/infrastructure/datasources/WebServices.dart';
+import 'package:cobble/infrastructure/datasources/web_services.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class RebbleSetupSuccess extends StatefulWidget {
